requirements.txt
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
python-dotenv==0.20.0
Werkzeug==2.2.2
gunicorn==20.1.0
sendgrid==6.9.7
sendgrid-python==0.1.1
waitress==2.1.2
